refactor for easier management of the reference file
