+++ lib/chat/toolChains/create_release_report/referenceReleaseReport.ts
@@ -1,17 +1,5 @@
-import { ToolChainItem } from "./toolChains";
-import { ModelMessage } from "ai";
-
-/**
- * Creates a reference message with the release report example
- */
-const getReleaseReportReferenceMessage = (): ModelMessage => {
-  return {
-    role: "user",
-    content: [
-      {
-        type: "text" as const,
-        text: `Here is an example release report for reference. Use this as a template for creating your own release reports:\n\n
-        Hi all,
+export const referenceReleaseReport = `
+Hi all,
 
 Midweek update on all things PinkPantheress and ‘Illegal’ below.
 
@@ -159,45 +147,4 @@ Thanks!
 Zak Boumlaki
 Head of Marketing | Warner Records  
 27 Wrights Lane, Kensington, London W8 5SW
-(0) +44 7984 662 662`,
-      },
-    ],
-  };
-};
-
-export const createReleaseReportToolChain: ToolChainItem[] = [
-  { toolName: "search_web", messages: [getReleaseReportReferenceMessage()] },
-  { toolName: "youtube_login", messages: [getReleaseReportReferenceMessage()] },
-  {
-    toolName: "get_youtube_channels",
-    messages: [getReleaseReportReferenceMessage()],
-  },
-  {
-    toolName: "get_youtube_channel_video_list",
-    messages: [getReleaseReportReferenceMessage()],
-  },
-  {
-    toolName: "get_spotify_artist_albums",
-    messages: [getReleaseReportReferenceMessage()],
-  },
-  { toolName: "search_web", messages: [getReleaseReportReferenceMessage()] },
-  {
-    toolName: "create_knowledge_base",
-    messages: [getReleaseReportReferenceMessage()],
-  },
-  {
-    toolName: "generate_txt_file",
-    system:
-      "Create a Release Report with the information you have gathered about the release across spotify, youtube and web search.",
-    messages: [getReleaseReportReferenceMessage()],
-  },
-  {
-    toolName: "update_account_info",
-    system:
-      "Attach the newly created release report to the artist's account info as a knowledge base.",
-  },
-  {
-    toolName: "send_email",
-    messages: [getReleaseReportReferenceMessage()],
-  },
-];
+(0) +44 7984 662 662`;
