+++ public/release-report-reference.txt
@@ -1,0 +1,149 @@
+Hi all,
+
+Midweek update on all things PinkPantheress and ‘Illegal’ below.
+
+‘Illegal’ breaks into 6 new markets via local Spotify Top-200, charting in 22-markets overall, and a sharp surge in the UK, US and Global charts, all within the Top-100.
+
+In the UK ‘Illegal’ is midweeking at #22, forecasting #19 for the end of this chart week. ‘Fancy That’ is slowly climbing the charts, midweeking #163 (+12%), week previous charting at #180.
+
+Streaming headlines:
+
+‘Illegal’ is up 53% this-week-to-date, across all platforms
+Spotify is up 58% Monday’s streams were up 15% DoD 
+Apple Music is up 54%, all streams still growing daily across all sources 
+Amazon Music is up 25%, radio and playlist stream driving daily increase
+
+Spotify UK Chart: #20 (+21) 
+Spotify Global Chart: #86 (+55) 
+Apple Music UK Chart: #39 (Flat)
+
+Global streaming headlines:
+
+Illegal 
+Total Global Streams: 50.6M (+8%) 
+Week Prior Global Streams: 9.7M (+61%)
+This Week Global Streams: 7.9M (+54%)
+
+Fancy That
+Total Global Streams: 172M (+3%)
+Last Week Global Streams: 17.5M (+31%)
+This Week Global Streams: 11.2M (+16%)
+
+Update below.
+
+TikTok Story So Far: 
+June 2nd:
+
+Sound 1, here - 471,000 Creations + 53,000 
+Sound 2, here - 100,000 Creations +18,000
+Sound 3, here - 15,900 Creations +2,800
+Sound 4, here - 6,700 Creations +3,700
+
+‘Illegal’ Charts: 
+July 2nd:
+
+Spotify:
+#20 United Kingdom +21
+#42 Estonia +1 
+#40 New Zealand +15
+#43 Australia +15
+#37 Latvia +30  
+#53 Lithuania +21
+#60 United States +41 
+#77 Iceland +33 
+#72 Ireland +45 
+#71 United Arab Emirates +52
+#85 Global +56 
+#107 Austria +22
+#110 Norway +31
+#130 Germany +23 
+#126 Netherlands +38
+#122 Switzerland +45
+#114 Canada +59 
+#118 Singapore +62
+#133 Bulgaria 
+#144 Belgium  
+#151 Poland  
+#167 Saudi Arabia  
+#198 Romania
+
+Apple Music:
+#6 Iceland
+#18 Bahrain
+#39 United Kingdom
+#68 Lebanon
+#70 Kuwait
+#82 Sweden
+#83 Netherlands
+#85 Lithuania
+#120 Finland
+#136 New Zealand
+#137 Poland
+#139 Ireland
+#151 Norway
+#190 Switzerland
+#197 Austria
+
+iTunes:
+#19 Egypt
+#20 Qatar
+#168 United Kingdom
+
+Shazam:
+#32 Belarus
+#35 Norway
+#38 Germany
+#46 Switzerland
+#49 Czech Republic
+#51 Australia
+#51 Singapore
+#52 Austria
+#55 United Kingdom
+#61 Israel
+#64 Ireland
+#64 Netherlands
+#66 Sweden
+#72 Belgium
+#75 Worldwide
+#77 Poland
+#79 Canada
+#79 Finland
+#80 France
+#89 United States
+#95 New Zealand
+#108 United Arab Emirates
+#114 Hungary
+#141 Greece
+#148 Denmark
+#163 Malaysia
+#166 Bulgaria
+#175 Saudi Arabia
+#177 Philippines
+#179 Ukraine
+
+Deezer:
+#57 Indonesia
+#57 Lebanon
+#73 Singapore
+
+‘Illegal’ Toolkit:
+Assets:
+
+‘Illegal’ Official Artwork: https://isaac.wmg.com/?w=NhQb4Fp6-2 
+'Illegal' DSP Images: https://isaac.wmg.com/?w=x_yQPkH971 
+‘Illegal’ Official Video: https://isaac.wmg.com/?w=NHFiiVqudB 
+'Illegal' Cutdowns: https://isaac.wmg.com/?w=j4v4Yu814a 
+'Illegal' Lyric Video: https://isaac.wmg.com/?w=-Kq7p_BfOX 
+‘Illegal’ Audio Ad: https://isaac.wmg.com/?w=5U2uWfklpi 
+‘Illegal’ Press Release: https://isaac.wmg.com/?w=jBa00Y3mHe 
+‘Illegal’ Spotify Inventory: https://isaac.wmg.com/?w=a0beYDZxoe 
+‘Illegal’ Pseudo Video: https://isaac.wmg.com/?w=SkQvHLvm4W 
+‘Illegal’ Victory Lap Asset: https://isaac.wmg.com/?w=BPyZFGCxJH
+
+Thanks!
+
+--
+Zak Boumlaki
+Head of Marketing | Warner Records  
+27 Wrights Lane, Kensington, London W8 5SW
+(0) +44 7984 662 662
